Tests update, bug fixes in office module, new docker image for pgadmin

+++ src/main/java/pl/edu/agh/rentableoffices/office/dto/CreateOfficeCommand.java
@@ -25,7 +25,7 @@ public class CreateOfficeCommand {
     private Integer maxTenants;
 
     @Positive
-    private Float area;
+    private Double area;
 
     @NotNull
     private AddressDto address;
