+++ src/main/java/pl/edu/agh/rentableoffices/office/dto/OfficeDto.java
@@ -20,7 +20,7 @@ public class OfficeDto {
     private Integer roomCount;
 
     @Positive
-    private Float area;
+    private Double area;
 
     @Embedded
     private AddressDto address;
